requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
email-validator==2.1.0
